statbot-intent-detector-evaluation/requirements.txt
GitPython==3.1.40
langdetect==1.0.9
nltk
numpy
psycopg2-binary
py3langid==0.2.2
pytablewriter==1.2.0
python-dotenv==1.0.0
tqdm==4.66.1
Unidecode==1.3.7
transformers==4.35.2